MOD1 9 My Solution
This was my solution to the "Exercise Card Component". I got it
to the best of my abilities. I could not figure out how to load
the image for Astro image processing from the json data. I eneded
up setteling with putting the images into the public folder so
that they could just be directly referenced in the json data.
I will watch for Chris Pennington's solution to see if he covers
this. Also I didn't take the time to set each article category
with a color and just mafe them all the same color for the sake
of time.